Add suspenses to the right places
