+++ src/graphql-hooks/chats-hooks.ts
@@ -29,6 +29,7 @@ const useChatAdded = () => {
 
     }
 
+    if (!chats) return
     if (chats.some(chat => chat.id === chatAdded.id)) return
 
     chats.push(chatAdded)
